refactor: move nplike finding to its own submodule
